Reorganize controller file structure
