+++ src/controllers/map-state.js
@@ -1,4 +1,4 @@
-import PerspectiveMercatorViewport from '../../viewports/web-mercator-viewport';
+import PerspectiveMercatorViewport from '../viewports/web-mercator-viewport';
 import assert from 'assert';
 
 // MAPBOX LIMITS
