--- src/controls/index.js
@@ -1,3 +1,0 @@
-// Deprecated exports, just to avoid breaking the 4.1 examples
-export {default as MapController} from '../react/controllers/map-controller';
-export {default as OrbitController} from '../react/controllers/orbit-controller';
