+++ src/react/controllers/map-controller.js
@@ -1,9 +1,9 @@
 import {PureComponent, createElement} from 'react';
 import PropTypes from 'prop-types';
 
-import EventManager from '../../controls/events/event-manager';
-import MapControls from '../../controls/map-controls/map-controls';
-import {MAPBOX_LIMITS} from '../../controls/map-controls/map-state';
+import EventManager from '../../controllers/events/event-manager';
+import MapControls from '../../controllers/map-controls';
+import {MAPBOX_LIMITS} from '../../controllers/map-state';
 import CURSOR from './cursors';
 
 const propTypes = {
